refactor: use `u32` for PID

+++ src/os/linux.rs
@@ -17,7 +17,7 @@ pub(crate) fn get_open_sockets() -> OpenSockets {
             let Ok(fds) = process.fd() else { continue };
             let Ok(stat) = process.stat() else { continue };
             let proc_name = stat.comm;
-            let proc_info = ProcessInfo::new(&proc_name, stat.pid);
+            let proc_info = ProcessInfo::new(&proc_name, stat.pid as u32);
             for fd in fds.filter_map(|res| res.ok()) {
                 if let FDTarget::Socket(inode) = fd.target {
                     inode_to_procname.insert(inode, proc_info.clone());
